Added test for RouteGuider middleware
